Deployment/requirements.txt
pandas==2.2.2
redis==4.5.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.31